Move maker's `ActorSystem` to `lib` as `Maker`
